LayoutTests/fast/box-shadow/hit-test-box-shadow-on-zero-height-clipping-container-2-expected.txt

Hovering over the blue shadow area under the black bar should not make green go to red, for the whole width

PASS: target element class is "first box"
